requirements.txt
Django==5.0.6
djangorestframework
djangorestframework-simplejwt
drf-spectacular
psycopg2